Light proofreading on Yoigoshi scripts
Proofread all scripts and fixed typos, errant spaces, comma splices, and some grammatical issues. Also standardized internal styles for numbers, time, and capitalization. (The issue of whether the names of generic speakers, like "Man in Black," should be headline- or sentence-cased remains unresolved, however.)

Suggested one small edit to the speaker name "Town Treasurer." In context, it seems like "Local Loan Shark" is more what's intended. Up to you!

Also wanted to note that in yoig_badend2.txt, there's a line where the the kanji for "river" is intended to appear. It's there in the script, but doesn't show up in-engine.