Update kage_003.txt
Change "Chief Inspector Minai" for "Inspector Minai" for instances with 南井警部. 警部 (keibu) used to refer to just plain Inspector before there was a restructuring of the police force in Japan in 2013 (after console arcs released?) where it came to refer specifically to the Chief Inspector. So in this case, Minai is a lower rank than a "Chief Inspector". Unless there's further plot stuff I'm missing, in which case lmk!