Translation corrections
Various translation corrections.
End of my Tokihogushi corrections.
The どちらが菩薩で孫悟空なのか、あとで思い知らせてやりますよ revision was made unsure of the full context. It appears to be referencing the Journey to the West, where the Buddha imprisoned Sun Wukong under a mountain for 500 years. If the original translation matches the context of the full story and was done for localization purposes, it should stay.