TL improvement & Correcting typos
line 1575 in 003
"予告" refers to line 1323

line 881 in 004
I think summer break of Hinamizawa school hasn't started yet

line 252&259 in 007
ムキムキ comes from 剥く, and ヌギヌギ is from 脱ぐ

line 1354 in 007
えっぐぅ is a variation of えぐい.

line 1556 in 008
I think 黒光りするヤツ is a cockroach. It seems Rika hates cockroaches.
顔を見せる is a idiom that means "show up" or "appear."

line 1686-1709 in 008
Both 良心 and 両親 are ryoushin. This is a pun of Japanese. So it might be better to use something like annotations.