Update koto_010.txt
There are some voice-only line flashbacks from script 006 in this file. Since the scripting does not seem to mention these voice files in any way, they should be implemented somehow. Below are between which lines the flashback lines play in this script, followed by which line it corresponds to in 006.

300-320:  1329, 1335

322-347: 1761, 1803, 1806

349-374: 3041, 3050, 3053, 3056, 3059

377-401: 3786, 3789, 3798, 3804