Translation
This concludes the Kotohogushi translation, as well as my part in it. I will leave the editing and everything else up to you.

Translation notes:
I chose to refer to Hanyuu when she's being described as a "demon" as an "it" instead of a "her" to both emphasize that this isn't the Hanyuu we know, and to help differentiate her from Ouka in the narration.

Lines 3279-3609 contain lines both verbatim and paraphrased from lines 1467-1592 of Matsuribayashi 11 (matsu_011). The verbatim lines had their TLs copied from the official TL of Matsuribayashi, while the paraphrased lines are modified versions of the same lines.

There are voice-only lines from Hanyuu and Ouka from lines 3784-3802 not featured in the scripting. Since these lines are in Ryuunish to begin with, I do not believe it is necessary to "translate" them.


Kotohogushi ED ending scroll TL:
...After this tragedy,
Ouka Furude was revered in Onigafuchi Village
as the savior who slayed the rampaging demon that attacked the village and killed many.

It is said that the feudal lord, fearing the fury of the demon's wrath would fall upon him,
made no further excess demands of the village.
Furthermore, he entrusted the right to rule the village to the Sonozaki and Kimiyoshi families,
and took up a nonintervention stance as the village functioned as an autonomous settlement.

As a result, the Three Families gained even more influence in the village,
and due to the divine wrath they'd experienced firsthand,
the villagers were all firmly bonded by their belief in Oyashiro-sama.

...And, to ensure demons would never rise again,
Ouka Furude offered up the precious sword that had felled the demon to Onigafuchi Swamp.
The branching blade, bared and hardened with demon blood,
came to later be known as the "Onigari-no-ryuuou".

Every year afterwards, a festival was started under the initiative of the Three Families,
named "Watanagashi", where offerings were drifted down the river to purify the slain demon's grudge.
Later, these offerings would be changed into something harmless and amicable in the form of futon cotton.
(However, no documents remain regarding who proposed this change)

...However, these are all merely conjectures that have been formulated based on a series of ancient documents.
Furthermore, the most prominent documents were lost to fire, 
and is debated amongst experts due to its heavy reliance on heresay.

Therefore, the true history remains unsolved to this day,
and rides on the results of future archaeological research...