TL improvement & official TL import (#539)
* TL improvement & official TL import

I'm going to continue to recheck Miotsukushi a little more.

line 71 & 823 in omot_10
imported from Tsumihoroboshi (tsum_011.txt)

line 1759 in omot_10
I think the word "お盆" made a misunderstanding. I'm certain that お盆 means a tray in this case. (お盆 also has a meaning of a Japanese traditional festival which takes place in August)

* Update Update/omot_010.txt

* Update Update/omot_010.txt

* Update Update/omot_010.txt

* Update Update/omot_010.txt

* Update Update/omot_010.txt

* Update Update/omot_010.txt

* Update Update/omot_010_2.txt

* Update Update/omot_010_4.txt

* Update Update/omot_010_4.txt

* Update Update/omot_010_4.txt

* Update Update/omot_010.txt

* Update Update/omot_010.txt

Co-authored-by: DoctorDiablo <doctordiablo@gmail.com>